docsbox/requirements.txt
#git+https://github.com/nvie/rq.git#egg=rq
#git+https://github.com/nvie/rq.git#egg=rq
#rq
redis==2.10.6
rq-scheduler==0.6.1
#rq-scheduler
rq==0.12.0
Flask==0.11.1
Flask-RESTful==0.3.5
Flask-RQ2==16.0.2
flask-env-settings==0.1.0
gunicorn==19.6.0
pylokit==0.8.1
python-libmagic==0.4.0
wand==0.4.3
ujson==1.35
nose==1.3.7
cffi==1.7.0
coverage==4.5
#pypdf3==1.0.1
pypdf3
